test: move test_save_certs() into own executable
Move the `test_save_certs()` test function from
`tests/voucher/test_request.c` to
`tests/voucher/generate_test_certs.c`.

This is because we will use this executable to generate some test files
used by other tests.
